Change handling of spring.config.location so it replaces defaults
Previously spring.config.name was handled in such a way that its value
would replace the defaults. By constrast, spring.config.location would
add to the defaults.

This commit updates the handling of spring.config.location so that it
replaces the defaults. This aligns its behaviour with
spring.config.name.

Closes gh-10595