Write META-INF before launcher classes when creating fat jars and wars
Previously, the launcher classes were written out as the first entries in a
fat jar or war. This broke an unspecified assumption [1] made by JarInputStream
that META-INF will always be written first.

This commit reworks BootJar and BootWar so that META-INF is written first, then
the launchers classes, followed by the application's classes and lastly jars of
its dependencies.

[1] https://bugs.openjdk.java.net/browse/JDK-8031748