Use alternative fallback mechanism for Tomcat's jar:war: URLs
Previously, opening a connection to a Tomcat-created jar:war: URL
would result in an attempt to fall back to the JDK's handler for jar
URLs as Boot's handler only deals with jar:file: URLs. This fall back
requires the use of reflection and, when running on Java 13, this
reflective access is illegal and a warning is logged.

This commit updates Boot's handler for jar URLs to use a different
approach when falling back. jar:war: URLs that contain */ (as those
created by Tomcat do) are handled specially by creating a URL from
the jar:war: URL's file (so a war: URL) and opening a connection to
it. This results in Tomcat's war: URL handling being used and a
WarURLConnection being created without trying to access
sun.net.www.protocol.jar.Handler using reflection.