Polish manifests of source and javadoc jars
Previously, the main jar file and the source and javadoc jar files all
had the same changes applied to their manifests.

The commit changes the Implementation-Title of source and javadoc jars
so that the title indicates that's what they are. Rather than using
the project's description as the title (as is done for the main jar),
the titles for source and javadoc jars will now be "Source for
${project.name}" and "Javadoc for ${project.name}" respectively.