Use unique names for property sources created from wildcard locations
Previously, all the property sources created from the same wildcard
location would have the same name. Each time a property source that is
equal to an existing property source is added, it replaces the existing
property source. Property source equality is name-based so this
resulted in the last property sources from the wildcard location
winning.

This commit updates StandardConfigDataLoader to use the resolved
Resource rather than the wildcard location in which it was discovered
in the name of the property source that it creates, ensuring that each
is property source from a wildcard location is uniquely named.