Fix constructor binding in a native image without -parameters
If a class is compiled without -parameters discovery of the names of
its constructor parameters falls back to
LocalVariableTableParameterNameDiscoverer. Its discovery process
requires the .class file to be available as a resource so that it can
be read using ASM. It also requires every method in the class to be
resolvable to an Executable. This commit adds the necessary resource
and reflection hints so that this works in a native image.