Introduce service connections that influence auto-config
This commit updates several auto-configuration classes to use an
optional ConnectionDetails bean as a source of configuration
settings for their connection to a remove service. When such
connection details are available, they take precedence over any
connection-related configuration properties. Configuration
properties that are not related to the connection itself, such
as properties that control the size and behavior of a connection
pool, are still used.

Close gh-

Co-Authored-By: Phillip Webb <pwebb@vmware.com>
Co-Authored-By: Mortitz Halbritter <mkammerer@vmware.com>